% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etamsdatap_de.R
\name{etamsdatap_de}
\alias{etamsdatap_de}
\title{measured confounder}
\usage{
etamsdatap_de(eta, ratio, changevariable, msdatap)
}
\arguments{
\item{eta}{\eqn{\eta} in "Sensitivity analysis.pdf"}

\item{ratio}{\eqn{\zeta} in "Sensitivity analysis.pdf"}

\item{changevariable}{the names of variables that used to create a measured confounder; ex: c("Q", "W")}

\item{msdatap}{the data only contain the exposure, mediators and outcome; ex: cbind(Y, Q, S, W)}
}
\value{
the measured confounder without the disturbance \eqn{\epsilon}
}
\description{
This function is for sensitivity analysis used in senstivityanalysis.txt
It will calculate the measured confounder without the disturbance\eqn{\epsilon}
reference: 'Sensitivity analysis.pdf' in 'ref' file folder
}
