% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_PSE_U.R
\name{mean_PSE_U}
\alias{mean_PSE_U}
\title{PSE with measured confounder add in the model}
\usage{
mean_PSE_U(
  eta_msdatap,
  msdata,
  confounders_ms,
  n_cate,
  sc,
  intv,
  intval,
  nb = 0,
  etasd,
  n_core = 1
)
}
\arguments{
\item{eta_msdatap}{the measured confounder without esp}

\item{msdata}{the orginal data without measured confounder}

\item{n_cate}{the number of categories of the outcome}

\item{sc}{the category of Y that users have "NO" interesting in}

\item{intv}{Number of intervention, only 4.}

\item{intval}{the value of exposure used in Intervention , Default is c(0,1).}

\item{nb}{Number of bootstrapping. Default is 0 (no bootstrapping applied).}

\item{etasd}{the stand deviation of \eqn{\eta}}

\item{n_core}{number of cores that will be used in parallel computing. default=1}

\item{confounder_ms}{the values of the confounder}
}
\description{
This function is for sensitvity analysis used in senstivityanalysis.txt
It will calculate the measured confounder without the disturbance(epsilon)
reference: 'Sensitivity analysis.pdf' in 'ref' file folder
}
