% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSE_0.R
\name{PSE_0}
\alias{PSE_0}
\title{This function will separate out the PSE of the orginal data from the output of function 'Ord_mediation_analysis_pal'.
This subfunction is used in sensitivityanalysis.R}
\usage{
PSE_0(PSElist, catecount, w_c, pathcount, dor)
}
\arguments{
\item{PSElist}{the output of function 'Ord_mediation_analysis_pal' with orginal data}

\item{catecount}{for counting the categories of outcome that user have interesting in}

\item{w_c}{category of Y that user have instercting in}

\item{pathcount}{fot counting the pathway}

\item{dor}{must be "D" or "R", for idenfity "difference" or "ratio" scale to be used}

\item{tiile}{The PSE for the orginal data}
}
\description{
This function will separate out the PSE of the orginal data from the output of function 'Ord_mediation_analysis_pal'.
This subfunction is used in sensitivityanalysis.R
}
